vcodec: CB2 modify prepare_get_addr return type [1/1]
PD#SWPL-96398

Problem:
dma buffer virtual address can not handle by kernel.

Solution:
Modify prepare_get_addr return type.

Verify:
t7c

+++ drivers/amvdec_ports/aml_vcodec_dec.c
@@ -3147,7 +3147,7 @@ int aml_uvm_buff_attach(struct vb2_buffer * vb)
 	return ret;
 }
 
-static int prepare_get_addr(struct dma_buf *dbuf, struct device	 *dev)
+static ulong prepare_get_addr(struct dma_buf *dbuf, struct device	 *dev)
 {
 	struct dma_buf_attachment *dba;
 	struct sg_table *sgt;
